+++ unicat_online_learning/src/main/java/com/unicat/onlinelearning/controller/ErrorHandler.java
@@ -4,7 +4,6 @@
  */
 package com.unicat.onlinelearning.controller;
 
-import com.unicat.onlinelearning.dao.UserDAO;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
@@ -12,22 +11,19 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-/**
- *
- * @author Duc Anh
- */
-@WebServlet("/delete")
-public class deleteUser extends HttpServlet{
+@WebServlet("/handle-error")
+public class ErrorHandler extends HttpServlet {
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        req.getRequestDispatcher("error_page.jsp").forward(req, resp);
+    }
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-    String id = req.getParameter("uid");
-    UserDAO dao = new UserDAO();
-    dao.deleteUser(id);
-    resp.sendRedirect("admin/manager/user");
-        
+        String string = "";
+        doPost(req, resp);
+        req.getRequestDispatcher("error-handler").forward(req, resp);
     }
-    
-    
-    
+
 }
