+++ unicat_online_learning/src/main/java/com/unicat/onlinelearning/controller/course/Course_Review_Comment.java
@@ -1,4 +1,4 @@
-package com.unicat.onlinelearning.controller;
+package com.unicat.onlinelearning.controller.course;
 
 import com.unicat.onlinelearning.dao.ReviewCommentsDAO;
 import com.unicat.onlinelearning.dao.ReviewDAO;
