+++ unicat_online_learning/src/main/java/com/unicat/onlinelearning/controller/course/Courses_Pagination.java
@@ -1,4 +1,4 @@
-package com.unicat.onlinelearning.controller;
+package com.unicat.onlinelearning.controller.course;
 
 import com.unicat.onlinelearning.dao.BlogDAO;
 import com.unicat.onlinelearning.dao.CategoryDAO;
