+++ unicat_online_learning/src/main/java/com/unicat/onlinelearning/controller/user/RestFB.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.unicat.onlinelearning.controller;
+package com.unicat.onlinelearning.controller.user;
 
 import java.io.IOException;
