+++ unicat_online_learning/src/main/java/com/unicat/onlinelearning/controller/user/deleteUser.java
@@ -1,0 +1,33 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package com.unicat.onlinelearning.controller.user;
+
+import com.unicat.onlinelearning.dao.UserDAO;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+/**
+ *
+ * @author Duc Anh
+ */
+@WebServlet("/delete")
+public class deleteUser extends HttpServlet{
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    String id = req.getParameter("uid");
+    UserDAO dao = new UserDAO();
+    dao.deleteUser(id);
+    resp.sendRedirect("admin/manager/user");
+        
+    }
+    
+    
+    
+}
