+++ unicat_online_learning/src/main/webapp/WEB-INF/web.xml
@@ -6,4 +6,12 @@
     <welcome-file-list>
         <welcome-file>home</welcome-file>
     </welcome-file-list>
+    <error-page>
+        <error-code>404</error-code>
+        <location>/handle-error</location>
+    </error-page>
+    <error-page>
+        <exception-type>java.io.IOException</exception-type>
+        <location>/handle-error</location>
+    </error-page>
 </web-app>
