+++ unicat_online_learning/src/main/webapp/styles/error_style.css
@@ -1,0 +1,54 @@
+/*
+Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/CascadeStyleSheet.css to edit this template
+*/
+/* 
+    Created on : Nov 5, 2022, 11:24:57 PM
+    Author     : Anhnt206
+*/
+
+
+/*======================
+    404 page
+=======================*/
+
+
+.page_404{
+    padding:40px 0;
+    background:#fff;
+    font-family: 'Arvo', serif;
+}
+
+.page_404  img{
+    width:100%;
+}
+
+.four_zero_four_bg{
+
+    background-image: url(https://cdn.dribbble.com/users/285475/screenshots/2083086/dribbble_1.gif);
+    height: 400px;
+    background-position: center;
+}
+
+
+.four_zero_four_bg h1{
+    font-size:80px;
+}
+
+.four_zero_four_bg h3{
+    font-size:80px;
+}
+
+.link_404{
+    color: #fff!important;
+    padding: 10px 20px;
+    background: #39ac31;
+    margin: 20px 0;
+    display: inline-block;
+    border: 0px;
+}
+.contant_box_404{
+    
+    margin-top:-50px;
+    text-align: center;
+}
